Make Watcom happy when comparing 64 bit vars with 32 bit constants:
parser.c(1033): Warning! W124: Comparison result always 1
parser.c(1035): Warning! W124: Comparison result always 1
./output/outobj.c(1142): Warning! W124: Comparison result always 0

Hand-picked from master commits 87646096a9c7afa39e581dbc236c5f77d39ba8b1
and e90a89abe47bfc1146f1a2a77409ce103789d9f8

+++ output/outobj.c
@@ -1139,7 +1139,7 @@ static void obj_out(int32_t segto, const void *data,
 	    ldata += size;
         }
 
-	if (size > UINT_MAX)
+	if (size > (uint64_t)UINT_MAX)
 	    size = 0;
 
 	switch ((unsigned int)size) {
